# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Streaming App API",
        "description": "API documentation for the streaming application including user authentication, subscriptions, chat, and payment features",
        "contact": {
            "name": "API Support",
            "email": "support@streamingapp.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/user/usersuperstar": {
            "get": {
                "tags": [
                    "User Superstars"
                ],
                "summary": "Get all superstars with detailed information",
                "description": "Fetch all superstars with username, DP image, total posts, followers count, subscription status, and cost",
                "operationId": "7fa650caf6c4f409b5c7ffab0ba9128e",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Superstars retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "username": {
                                                                "type": "string",
                                                                "example": "johndoe123"
                                                            },
                                                            "dp_image": {
                                                                "type": "string",
                                                                "example": "https://example.com/images/user123.jpg"
                                                            },
                                                            "total_posts": {
                                                                "type": "integer",
                                                                "example": 25
                                                            },
                                                            "followers": {
                                                                "type": "integer",
                                                                "example": 1500
                                                            },
                                                            "is_subscribed": {
                                                                "type": "boolean",
                                                                "example": true
                                                            },
                                                            "cost": {
                                                                "type": "number",
                                                                "format": "float",
                                                                "example": 50
                                                            },
                                                            "display_name": {
                                                                "type": "string",
                                                                "example": "John Doe"
                                                            },
                                                            "bio": {
                                                                "type": "string",
                                                                "example": "Professional superstar"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 15
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 42
                                                },
                                                "from": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "to": {
                                                    "type": "integer",
                                                    "example": 15
                                                },
                                                "has_more_pages": {
                                                    "type": "boolean",
                                                    "example": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format: Bearer {token}",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}